Create constructor BigDataAirportHotelBaselStartFrame.
We create the constructor of the class and within the constructor the
respective instances.

note that the instances of the buttons(MyButton) display some error
because i have not yet created the constructor with parameters of this
class(MyButton).

Class MyButton should receive one String with the path where the image
is located to use with our custom buttons.