BigDataAirportHotelBaselStartFrame call setupFrame
We create the setupFrame Method. Inside we setTitle of the JFrame, the
Location, and the setDefaultCloseOperation(DISPOSE_ON_CLOSE) very
important.

We set Incon to the JFrame

instantiate the class PanelWithBackgroundOption and we indicate the path
to the image that will have for the background.

setContentPane(Our Class PanelWithBackgroundOption with the image).

We call a Method named addElementsToPanel. method that deals with
putting the elements in the panel.


In de Class PanelWithBackgroundOption i create the Method setImage. this
method receives an argument with the image path as String variable.