LogicModelStartFrame add some variables, method and a new Class
New Class JTextFieldLimit to limit the number of characters that can be
entered inside a JTextField.

By the Class LogicModelStartFrame i have create some new variable. This
time, all the variables to be able to set the path of the local
database.

We have a custom dialog box, images for the dialog box and buttons add
ActionListener and KeyListener for the events.


This class LogicModelStartFrame will check if the configuration file
exists. If it does not exist, we start in the first instance to
establish the path of the database and then the administrator user will
be created.

In case the user does not set a path for the database, the application
closes.