new method to create administrator user...
I have created the necessary variables for the method createAdminUser.

The method will display a custom dialog box(JDialog). GridbagLayout will
be used in the custom dialog box.

We have an array for the buttons(okButtonAdmin to add the new Admin User
and abbrechenJButton to cancel and exit the program).

In case No Admin User was created, Application exit.

We have not yet put the buttons to listen and we have not called the
method that verifies the existence of the configuration file.