LogicModelStartFrame new Methods was created and Controller Class...
We have the Dialog Input Box JButtons added ActionListener and
KeyListener now we have configured in the Controller Class
BigDataAHBStartFrameController inside the Method windowOpened the call
to the Method checkConfigurationFileProperties located in
LogicModelStartFrame.

This Method checkConfigurationFoleProperties will be called when the
Window is opened.


LogicModelStartFrame checkConfigurationFileProperties is going to check
if Configuration File exist. in Case does not exist we call method
setDataBasePath.

Method setDataBasePath does his Job and it will be called Method
createAdminUser.

We have create the method createAdminUser for creating Administrator
users. This method will be automatically called the first time we run
our application.

We validate the Admin Entry, checkPasswordAdminEntry and
checkAbkuerzungMAAdminEntry.

New Methods checkPasswordAdminEntry and checkAbkuerzungMAAdminEntry was
created.


Inside the method checkAbkuerzungMAAdminEntry If everything went
satisfactorily, inside one block if will be checked if userName,
password and abkuerzungMa is not "".

Then we will have to create a method that encrypts the data before
saving it to the configuration file.