Add WindowListener to BigDataAirportHotelBaselStartFrame...
I created a Constructor Method for the Class LogicModelStartFrame.

Initialize the Variables for the Custom Input Dialog box.

Initialize the Variable for the ImageIcon(imgUserRequest)

Initialize the JTextField with 20 Columns.

Initialize the passwordField with also 20 Columns.

Initialize GridBagLayout gbl and GridBagConstraints gbc.

With all these changes we have solved the problem of
NullPointerException. (I hate this NullPointerException LOL).