New method encryptDataUser by the class LogicModelStartFrame.
New Variables to set the Properties values for the Property File.

Add Instance DataEncryption.

Inside the Method checkAbkuerzungMAAAdminEntry it will be call the
Method encryptDataUser. to this method we pass the parameter
privilegePropertieValue. is the privilege for the Administrator User.

Inside the Method encryptDataUser, i set the value for the different
variables to be save in the Property File. After that we encrypt the
values and at the same time we give new values to the variables.

For the moment we print just the results. Later we need a new class for
writing the Properties. Class that use Properties class.