New Class PropertiesWriter to write properties to a File
with this class we are going to write the properties of the
configuration file.

First check if the property file exists. If the file exists, it opens to
add a new property to the file. If the file does not exist, it will be
created and the properties will be written to the File and save.