New Instance Class UserAHB and more...
Initialize UserAHB Class.

Method setLoginValue check if the userName and Password is not "".

setLoginValue set the value of the properties

setLoginValue call to encrypt the data.

New Class LoginDataUser

setLoginValue set the values of the LoginDataUser

BigDataAHBStartFrameController call inside method actionPerformed and
keypressed the setLoginValue Method from the logic class.