i removed everything that has to do the Focus.
Every methods that has something to do with setting the focus was
removed.

It will be modify in the next commit.

Also it was removed the FocusListener from the parkingButton inside the
Controller class.

Removed all the String variables userName, password, abbkuerzungMA,
privilege, urlDatabase from PropertiesReader Class 'cause we don't need.

in the controller class I removed the instruction to instantiate again
the logic class and the start class. These classes have been
instantiated in the runner class, where the main method is located.