Class AHBParkingController bigDataStartFrameController removed from..
Class and Constructor.

Class AHBParking We eliminate the implementations of WindowsListener,
KeyListener, ActionListener, FocusListener and implements them in the
AHBParkingController class.

AHBParking new setupFrame Method. inside the method we create a call to
the closeParking method.

closeParking Method addWindowsListener as parameter new WindowAdapter
and the windowClosing method inside call confirmClose Method. Used when
the user press the cross to close the JFrame.

Inside the AHBParkingController add Action and KeyListener to
exitParking and noExitParking JButton belonging to the custom dialog box
to close the parking JFrame.

added instructions for buttons exitParking and noExitParking inside the
keyPressed Method by the Parking Controller Class.

Also added instructions inside the actionPerformed Method.

And windowClosing Method by the AHBParkingController Class call
ahbParking.dispose.

added instructions when user presses F2 key to call the AHBParking
JFrame by the BigDataAHBStartFrameController Class.

BigDataAHBRunner call bigDataAHBStartFrameController.setMyIcon where we
check if we are running this application from a Apple Computer or not
and we set the Icon for our application.

Changes in methods keyPressed and actionPerformed
BigDataAHBStartFrameController