Some changes in classes and new classes for login.
AHBParkingController Class: Add ActionListener for all the MenuButtons.

AHBParkingController add add instructions in the actionPerformed method
for the different MyButtons.

BigDataAHBRunner: we pass as argument two new
parameters(bigDataAirportHotelBaselStartFrame and userAHB. It could be
changed i am working on it.

Changes in the LostFocus parkingButton in
BigDataAHBStartFrameController.

Remove setMyIcon Method from BigDataAirportHotelBaselStartFrame.

New Class LogicModelLogin for the login part of the Application.

I am making a new Login GUI.

Class LogicModelParking: the constructor receives two
parameters(bigDataAirportHotelBaselStartFrame, UserAHB).


Class LogicModelStartFrame significant new changes for this class.
Import Class LoginController, BigDataAirportHotelBaselStartFrame and
LoingUsser class.

Class LogicModelStartFrame removed all instructions from the custom
login dialog inside loginUser Method.

We call inside the loginUser Method invokeLater and inside the run
Method we instantiate the logicModelLogin class. at the moment this
class receives two parameters. This may change in the future..

Instantiate also the LoginController Class, we center the JFrame to the
screen and make it visible.

We still have inside the class LogicModelStartFrame the setLoginValue
method, but this method may change it for the LogicModelLogin class. i'm
still working on this part of the algorithm.

New Class LoginController. Constructor receive 3 parameters, LoginUser,
UserAHB, BigDataAirportHotelBaselStartFrame. this class is called to be
instantiated from the LogicModelStartFrame class. Takes charge of the
logic of the LoginUser class.

LoginUser is a new Class. this class is also instantiated from the
LogicModelStartFrame class. If the configuration file exists, the login
method is called, and when calling the login method, the LoginUser class
is instantiated and overlaps the
BigDataAirportHotelBaselStartFrame(JFrame) class.

The BigDataAirportHotelBaselStartFrame Class is blocked by the LoginUser
class that extends from JDialog. the BigDataAirportHotelBaselStartFrame
class will be locked until the user performs an action. If the user
presses cancel the entire application is closed. If the user presses
login after entering the data, the method to set the login data entered
by the user will be called.

The LoginUser class has a method named formComponentMoved used to
re-center the LoginUser(JDialog class) to the center of screen. In case
the user tries to drag the window, It will come back to the center of
the screen. To make this possible we add to (this) class
ComponentLisstener(new ComponentAdapter) and inside the componentMove
Method we call the formComponentMoved Method.

In the next commit i am going to work with the setLoginValues and check
the information entered by the user in the LoginUser(JDialog) class.