Login new changes.
BigDataAHBStartFrameController: modifying the call to the
logOutApplication method. This Method is now called inside the
logicModelStartFrame class and not anymore inside the
BigDataAHBStartFrameControlle.

Removed the logoutApplication method from
BigDataAHBStartFrameController.

REmoved every Parameters from the abstract method(logoutApplication)
inside the abstract class LogicModel.

Removed String variable privilegeUser from LogicModelLogin class, and
also the value given to this variable within the setLoginValue method.

Removed getter Method getPrivilegeUser from the LogicModelLogin class.

removed the initialization of the variable userAHB inside the
LogicModelStartFrame class.

important changes within the loginUser method: we add a conditional that
evaluates whether logicModelLogin has null value inside this method by
the LogicModelStartFrame class. In case it is null, we invoke a new
thread and create the instances, in case the value is not null, we
simply make the class visible, before i give values to the variables of
the userAHB class, calling the Method setUserName, setPrivilege,
value("").

removed getUserAHB by the class LogicModelStartFrame, not need it.

Class LoginController keyPressed method: change we have removed the
initialization of the value of the variable userAHB.

the instruction for the value of the variable privilegeUser is now
inside the try catch block because we call direct to
dataEncryptio.decrypData(this.userAHB.getPrivilege()). The same inside
actionPerformed method.

README.md new src for the img tag(Screenshot).