New classes for creating the parking table.
TableParkingUtilities for creating the value of each columns that we are
going to have in the Parking Table.

CellTableManager class: Class to manage the table and the eventss
performed on it.

ParkingTableModel: Class to build the model of our Parking Table.

TableHeaderManage: Class to customize the table headers, to de3fine the
color that columns will have etc.


TableParkingUtilities: Class to create the value of each columns in the
ParkingTable.

ParkingTable: Class for the Parking Table.

At the moment the table is empty. We just created the element and
included it in the center panel.


news by AHBParking: new Instance JTable parkingTable and JScrollPane for
the ParkingTable. add(ParkingTable).