AHBParkingController all for checkParkingDataBase.
LogicModelParking New DataEncryption variable, String urlDataBase to
store the Path where the DataBase should be saved, new LocalDateTime to
use for the name of the Database, dbName to store the name of the
DataBase, File urlDataBaseFile for checking if the DataBase exists.

New Method checkParkingDataBase and checkExistsDataBase Method.

removed some comments from LogicModelStartFrame.