New Class for DataBase Connection Motor derby EmbeddedDriver
Class ConnectionClass in the dao package.

the derby database engine is used with an encrypted database and an
encryption key present within the class.

New Class DaoException for the Exception.

New DAOParking interface.

New DaoParkingImpl extending of the ConnectionClass class and implements
DAOParking interface to implements all the methods of the interface.

It will check if the database exists, if the database does not exist it
will be created, then it will check if the Parking table exists, if it
does not exist it will also be created. Then we will be ready to start
writing data in our parking table within the database. of course we will
also be ready to request the data.