New Classes: NewParking, NewParkingController, LogicModelNewParking.
BigDataAHBStartFrameController add instruction in keypressed Method: new
Instance ParkingReservation and in the Instance of LogicModelParking we
pass this new parameter. The same in the actionPerformed Method.

New Picture for The NewParking GUI.

DaoParkingImpl new Variables for calling the NewParking GUI where we
create the new Parking Reservation. New Parameters by the constructor.

By DaoParkingImpl it will be check if we have any Parking reservation in
the DataBase(Table Parking). In case we don't have any reservation it
will be called the addNewParkingReservation Method and this take charge
of creating a new Instances LogicModelNewParking and
NewParkingController, after that it call newParking to visible.