DAOParking  new Method createNewParkingReservation.
by DAOParking createNewParkingReservation will be used for creating the
new Parking Reservation and not the addNewParkingReservation as before.


LogicModelNewParking many new changes i am still working on this class.

New variables closingNewParkingReservation use to know if the user try
to close the newParking Reservation GUI during enter the Data.

errorDateDepartureMessageDelivered and
errorDateArrivalMeswssageDelivered to play with the lostFocus between
both dates and avoid displaying error not needed.


NewParkingController inplements MousListener used to play with the
lostfocus by arrival date and departure date.

there other small changing it would be commented in next commit. I have
to go to other Job. LOL.