New changes to existing classes.
DAOParking removed parameter long in addNewParkingReservation Method.
Also the same for DaoParkingImpl class inside the implemented
addNewParkingReservation Method.


DAOParking addNewParkingReservation receive several new parameters.
parameters that we will use to store the data received in the database.

DaoParkingImpl addNewParkingReservation we also pass the parameters to
the implemented Method.


LogicModelNewParking inside the addNewParkingReservationToDataBase
method we call the new Instance daoParking to also call
addNewParkingReservation method with all the parameters we need to pass.
i'm still working on this method, it will be for later commits