Changes to add new parking reservation in some classes.
DaoParking receive now only one parameter(ParkingReservation).

DaoParkingImpl receive also only the ParkingReservation parameter.

LogicModelNewParking addNewParkingReservationToDataBase method create a
new instance of Parkingreservation and set all the values.

LogicModelNewParking call daoParking.addNewParkingReservation passing
the ParkingReservation parameter.

DaoParkingImpl removed ParkingReservation parameter by the constructor
method.

All the calls of daoParking removed also the ParkingReservation
parameter.

DAOParking removed createTableParking method and also the same method
was removed by DaoParkingImpl we don't need it anymore because every
time when we call for Parking GUI it will be checked if we have a
Database and if we have a table if not it will be created inside the
displayListParking.



Very Important:
many changes within the displayListParking Method: swingWorker for
Threads etc.

ConnectionClass method getConnection was created by after i comment it.

DaoParkingImpl class: we have changed the displayListParking method so
that it does not return anything.

The same as above for the DaoParking Class.

By DaoParkingImpl was removed the parameter Long lenghtParkingTable and
also it was removed inside each call to the Method
createNewParkingReservation.