New changes for displaying Parking Reservations on the Table.
AHBParking JTable parkingTalbe is now public so we can access to it.

CellTableManager new important changes: We create a SimpleDateFormat for
the Date, and we evaluate if the value is Date, Integer, or Double and
depending on the value we modify the value of the variable value to
display in the JTable Cell the text.


DaoParkingImpl imports DefaultTableModel, CellTableManager and
TableParkingUtilities for doing the job who displays the data in the
JTable.


DaoParkingImpl removed the instruction of System.out.println.

DaoParkingImpl new Array Object[] with all the data we need to display
in the JTable.

We get all the data for each column and cells and we the help of the
TableModel we add the new row.

Is nice to see some results but we continue working on this project.