LogicModelNewParking removed isBefore(nowLocalDate) instruction.
we have allowed the exit date to be earlier than today's date so that we
can edit old dates in the database.

This will help us to be able to edit delayed parking reservations that
were perhaps forgotten to enter in the database.

What if it is not going to be allowed as already specified, is that the
departure date is prior to the arrival date.