New classes for configuring local database directories and more...
ConfigurationDirectory GUI Class for configuring directories.

ConfigurationDirectoryClontroller Class.

LogicModelConfigurationDirectory.

AdminCreator Class to create administrator user accounts.

AdminCreatorController, LogicAdminCreator Class.

We added a new library(applejavaextensions-1.4.jar) so we can use this
to set the JFrame icon in the case we run the application in macOs
Operating System and if is not the case the application does not hang,
because in case we run the application on windows and the java virtual
machine does not contain the Application class belonging to the original
package com.apple.eawt.Application. As we have included this library,
then our application will not hang.

New Pictures for the JButtons(MyButton).

BigDataAHBRunner new variable static SetIconOperatingSystem, Class that
will take care of putting the icon to the JFrame depending on the
operating system we are running.

BigDataAHBRunner: We evaluate which operating system we are running and
depending on the operating system we add some system properties for
macOS.


BigDAtaAHBStartFrameController we import ComponentEvent and
ComponentListener to center the JFrame when the user try to move the
JFrame from place. The JFrame will return to the center of Screen.

New Class CenterComponentMoved to center our components according to the
component they have below.

DaoParkingImpl new deleteDirectory Method. I still to work on this idea.

New Classes LogicAdminCreator etc.

New Class OperatingSystemCheck to check the Operating System we are
running.

PropertiesReader new Variable urlDataBaseBackupPropertieName(String) for
the DataBase Backup Url.

New Class SetIconOperatingSystem for setting the icon to the JFrame
depending of the Operating System we are running.

UserAHB included now variables to save and the the DataBaseBackup url.


Some important changes in the LogicModelStartFrame Class to work with
the Configuration Directory and the AdminCreator.