DaoFactory add by one of the Constructor the instruction to call the
Method createDataBase();

DaoFactory by the createTable Method we add case "FUNDSACHEN" - Lost and
found. added the SQL instruction to create the FUNSACHEN table with the
columns we need.

DaoFundSachenImpl class was enriched with necessary variables. also was
created the tableChecked boolean variable to play with the Loading
Class.

By the checkTableFundsachen method we write the instructions needed. We
use one SwingWorker for the Loading Object with the JProgressBar. We
check if the FUNDSACHEN table exists and if do not exists we call the
daoFactory object exactly the createTable Method passing as argument the
FUNDSACHEN String for creating the FUNDSACHEN Table in the Database.

setURLToConnectCurrentDataBase method do the same as by Parking.

by the displayListFundsachen Method we only display in console for the
moment a message just to indicate is working but we are going to modify
this method in the next commits.

Also by the displayListFunsachen method it will be called the
dataBaseGUI to visible.

The other 3 Methods getDBName, setDerbyURL and getDerbyURL are the
similar like by Parking.

DaoParkingImpl: we removed some unnecessary comments.

DaoParkingImpl: by the checkTableParking method we use now one
SwingWorker to display the Loading Object with the JProgressBar.

Fundgegenstad Class we create a new String variable rueckGabe to save
the information when the Lost and found item was picked up by the owner
or someone sent by the owner. We also created the Methods getter and
setter.

LogicModel: we removed som unnecessary comments by the checkExistsDataBase Method.