2 New images btn_fundsachen2.png and btnfunsachenplus.png.
DataBaseGUI: new JButton object btnNewFundsachen. This JButton was
initialized with the appropriate image.

DataBaseGUI: case FUNDSACHEN new TopPanelFundSachen was called with the
new Argument. The new JButton for Lost And Found to be add.

TopPanelFundSachen: new JButton Variable and the constructor has a new parameter. It was add to the Panel the new JButton.