DaoFactory new Instructions by createTable Method for Lost and Founds.
DaoFundsachenImpl many new Instructions: new Instance Fundgegenstand,
DataEncryption.

DaoFundsachenImpl add instructions inside addNewFundsachen Method.

DaoFundsachenImpl add instructions inside displayListFundsachen for
displaying the Lost And Found items got it from the Database and Table
FUNDSACHEN.

Fundgegenstand: wrote correctly the Method setKisteName(String
kisteName).

FundsachenTable class: add instructions to use ToolTipText in the JTable
very important when the content in one Cell is too long. I wrote the
same instruction for the ParkingTable.

LogicModelFundSachen DataBaseGUI is now static.

LogicModelNewFundsachen: I created some variables to save the new
Information for Lost and Found Items.

LogicModelNewFundsachen: new Instruction by the checkDateFormat Method
for checking if the date that the user entered is in the good expected
format. We have also checkKistenNummer Method and some other
instructions added for adding new Lost And Found items to the Database.
Or better to say to call the DAO Class for adding the information.

LogicModelParking: checkParkingDataBase Method removed is not need it to
call from this class anymore.

NewFundsachenController: We use now UIManager class to reset some
elements of the NewFundsachen Class.

NewFundsachenController: some Listener was added to the JButtons and
other Components.

NewFundsachenController: some instruction was added by the
actionPerformed and keyPressed Method.

NewFundsachenController: some new instructions was added to the
focusGained Method so we can reset the Border value of the Components by
NewFundsachen Class.

NewFundsachenController: implements also the ItemListener interface so we can get the value changed by the JComboBox by the NewFundsachen Class. For that we use the itemStateChanged Method Override.