LogicModelNewFundsachen: New Instruction by checkAllEntries Method.
LogicModelNewFundsachen: by checkAllEntries Method we call
newFundsachen.dispose Method to close the NewFundsachen GUI before we
create the new Instance of DAOFundsachen and call to addNewFundsachen.
dispose() will be called if all the entries are OK as expected.

On the Contrary i removed the call newFundsachen.dispose by the
NewFundsachenController Class because we are going to check first if
every Entries are OK and then close the GUI.