LogicModelNewFundsachen: removed and add some new variables to check.
LogicModelNewFundsachen: entryCompleted boolean was removed ->
new boolean variables for better checking if all entries was entered by
the user. So we can avoid the bugs saving in database even it was some
JTextField not fill out.

NewFundsachenController: I add new instruction by itemsStateChange so when select it reset the kisteNameJTextFiel Border.