Edit JTable Cells, Add new Data in Row, Update the Data in Database...
DAOFundsachen new Method(updateFundsachen) to update the modified data.

DaoFundsachenImpl: some new code instructions. displayFundsachen Method:
new int variable to display the id for each row in the Table from the
Database. We also set the value id in this method calling the
fundgegenstand Object. Array Object row received a new variable int for
the id calling the chunk.getId Method.

DaoFundsachenImpl: new Override Method updateFundsachen for updating the
Data to the Database.

DataBaseGUIController: import TableModelEvent, TableModelListener,
TableModel because we implements also the TableModelListener interface.

DataBaseGUIController: each JTable depending if which is visible add the
TableModelListener to interact with the modified data in the JTable.

DataBaseGUIController: override tableChanged(TableModelEvent) so we can
get to work for updating the data in the Database when in the JTable
data was modified.

Fundgegenstand class has new int id variable to get the id from the
Database an to set the id number for updating the data for example.

FundsachenTable: New Column ID was added to the JTable to be displayed.
It will be displayed the id from each elements in the Table FUNDSACHEN
from the DataBase. it will be created this column as number and also not
editable.

LogicModelFundSachen new Method(updateFundsachen). This Method will
prepared all the Data and setting to the fundgegenstand Object to send
this as argument to the updateFundsachen Method by the DaoFundsachenImpl
object.

TableFundsachenUtilities has a new int variable for the id.