JComboBox will be used by JTable(Fundsachen) as DefaultCellEditor.
DataBaseGUI new variables JComboBox and the String Array with the
choices to use as DefaultCellEditor by JTable.

DataBaseGUIController TableModel now declared by the Class so we can
access from any Method to this variable. By the tableChanged we give the
value to the model calling the TableModelEvent parameter and the Method
getSource.

LogicModelFundSachen: Some new changes by the updateFundsachen Method:

	. kisteNummer is initialized to 0;

	. We use one try catch block to catch the NumberFormatException;

	. Inside the catch block i evaluate the value KisteN with one switch
	  conditional to give a new value and depending of the value of KisteN,
	  I give a new value Number as String and the JTable Column 5 setValue
	  the new value.

	  . The rest of the code is the same as before and the values will be
continue updated in the Database calling the Dao object and the
corresponding Method.