CellTableManager: new configuration for cells with important conditions
CellTableManager: When type="important" we are going to have the text
Color in Red.

DaoFundsachenImpl add the type "important" to the Column RUECKGABE. The
text color in this cell by the JTable will be displayed in red.

FundsachenTable add the type "important" to the Column RUECKGABE. The
text color in this cell by the JTable will be displayed in red.