Some new Codes to search in DataBase Table Lost And Founds
DAOFundsachen: new Method searchByDateFundsachen

DaoFundsachenImpl: implements the searchByDateFundsachen Method.

DataBaseGUI: new Elements: RoundJTextField for search Text,
JComboBox<String> elements to select by what we are going to search in
database. We also initialize the new elements in this class.

DataBaseGUI: By adding the TopPanelFundSachen element to the GUI we call
the Constructor now with this two new elements.

DataBaseGUIController: add the Listener for those new Elements. We have
also the corresponding instruction to call the Method
searchResultsInDataBase in LogicModelFundSachen Class.

LogicModelFundSachen Class: some new elements for displaying error
message in case the DateFormat is wrong when we search by date. new
Methods searchResultsInDataBase and checkDateFormat

TopPanelFundSachen: we removed String[] choices. The choices for the
JComboBox is now created by the DataBaseGUI Class to initialize the
elements and so we can send them to the Constructor TopPanelFundSachen.

In the next commits i will continue working by searching in the data
base by date, Dao Pattern and continue by displaying the results.