New JButton MyButton to reload the DataBase.
DataBaseGUI: new JButton to reload the DataBase.

DataBaseGUIController add KeyListener to the reloadDbJButton and
actionPerformed Method call to display again the Results from the
DataBase again.

New home JButton with F1 KeyCode Text not yet the instruction for that.

TopPanelFundSachen add to the Panel Class the new JButton to reload the
DataBase.