New Methods and instructions to reload Lost And Found from DB.
DAOFundsachen: new Method reloadFundsachenData to reload the Lost and
found data and displaying in the the JTable by the GUI.

DaoFundsachenImpl override the Method reloadFundsachenData.

DataBaseGUIController: actionPerformed Method call new instruction to
call the realodFundsachen Method by the logicModelFundSachen Class.

LogicModelFundSachen: new Method reloadFundsachen. In this method it
will be initialized the DAO Object to call the Method to reload by
DaoFundsachenImpl.

btn_home.png was removed from Project.