New Design JButtons NewFundsachen MyButton and more...
New images for the button save and cancel for the NewFundsachen Class.

DataBaseGUI setIconImage in case we are going to work under the Windows
Operating System.

DataBaseGUIController new Parameter LogicModelStartFrame by the
constructor Method to access to this class.

DataBasseGUIController by the keyPressed Method new instructions to
access to the different GUIS by Key combination.

Fundsachen Class removed not needed any more. All Database JTable will
be displayed by DataBaseGUI Class.

LogicModel: by displayParking and displayFundSachen Method initialized
LogicModelStartFrame Class to send it to the DataBaseGUIControlle
Constructor as argument when we create a new Instance of this class.

LogicModelNewFundsachen: removed System.out.println instruction.

MyButton: Not any change at all.

NewFundsachen: saveFoundSachen and cancelFoundSachen now is declared
direct as MyButton and not as JButton. New JPanel infoJPanel and new
JLabel infoJLabel to use as JFrame Title so we call setUnderorated(true)
for this JFrame Class. panelWithBackgroundOption setLayout as
BorderLasout, so we add first the Title and next the other
GridBagLayouts with their added elements.

NewFundsachenController add FocusListener to the saveFoundSachen and
cancelFoundSachen so when they get the Focus and lost the Focus the
Opacity will be changed and User can know he is over the MyButton now
and know he could press Enter Key.