logout Application bugs fixed...
BigDataAHBStartFrameController: changes to call the logoutApplication
Method. The Method do not receive any parameter any more. It will be
also dispose the bigDataAirportHotelBaselStartFrame instance before
calling the logoutApplication Method.

By the DataBaseGUIController we make the same operation.

LogicModel removed LoginUser and LogicModelLogin instances.
logoutApplication Method do not have any parameter anymore. By
logoutApplication Method instructions we create one instance of
SetIconOperationgSystem to set the Icon depending the Operating System
we are running the Application because we are desposing the main GUI
BigDataAirportHotelBaselStartFrame.

LogicModel: also by the logoutApplication Method we create the
BigDataAirportHotelBaselStartFrame, the LogicModelStartFrame instances
and we call for a new BigDAtaAHBStartFrameController passing the
expected parameters. After that we call setIconOperationgSystem instance
to call the Method setIconJFrame to set the Icon to our Application.
Then we make visible the bigDataAirportHotelBaselStartFrame GUI JFrame.

LoginUser: by cancel close the application we call the instruction to requestFocus to userLoginJTextField.