Add ActionListener to the Buttons by DataBaseGUI and Database Search...
DataBaseGUIController add ActionListener to the btnParking and btnLogout
Buttons. Also actionPerformed Method add new instructions for this
buttons when click on it.

DaoFundsachenImpl: new changes by the searByLostAndFound Method. Now we
search Like statement.

DaoFundsachenImpl: The name of the Methods was correctly modified also the same for the call of those methods from the LogicModelFundSachen class and the Declaration by DAOFundsachen Class.