Buttons activated to load different Menus in Case Parking Data loaded..
LogicModelDataBaseGUI removed String guiCaller variable also the String
guiCaller parameter from the Constructor. I also removed the
getGUICaller Method is not needed anymore.

LogicModel Class: removed String guiCaller variable from the Class, not
needed any more.

LogicModel: by displayParking and displayFundsachen Methods we add some
instructions to check if bigDataAirportHotelBaselStartFrame is not null
and if its visible to dispose before we display the Loading GUI and
displaying the data in the DataBaseGIU class.

LogicModel: also inside the displayParking and displayFundsachen we
removed the second argument when we create a new instance of
LogicModelDAtaBaseGUI.

LogicModel: Very important change. by the checkExistDataBase Method i removed the switch block that has checked the guiCaller variable value. now Loading will received always bigDataAirportHotelBaselStartFrame, true just to block the bigDataAirportHotelBaselStartFrame in background, but anyway we are going to close any GUI in background and run first the Loading GUI and then display the DataBaseGUI with the Data we want to display in corresponding JTable. So we avoid the user game with the Buttons and any way is only a short time to display. Never trust the user.