Add new Elements to the JPanel TopPanelParking Class...
TopPanelParking: imported RoundJTextField, SelectCustomJComboBox,
SimpleRoundBorder Classes.

TopPanelParking: new JPanel containerSearchSelect.

TopPanelParking: new Parameters by the constructor: RoundJTextField,
JComboBox<String>, JButton for reloadDB.

TopPanelParking: initialize the new elements and added to the GUI.

TopPanelFundSachen: modified the PreferredSize Dimension for the
searchJComboBox element.

DataBaseGUI: removed the declaration initialization of
choicesSearchFundsachen Array and searchJComboBox by the Constructor.

DataBaseGUI: Depending if PARKING or FUNDSACHEN it will initialize the choicesSearchFundSachen array values so we can have the correct values for PARKING or for FUNDSACHEN to search after the selected item value in the corresponding Database-Table.