ArrayIndexOutOfBoundsException Bugs fixed by the Method updateFundsachen
I decided to catch the ArrayIndexOutOfBoundsException before store the
value to the variables inside the updateFundsachen Method because i
become this exception when i selected a row and after i try to find some
entries in Database going to the search box and pressed enter.

The Row in JTable still selected and thats why i became this
ArrayIndexOutOfBoundsException.

The updateFundsachen Method will called from DataBaseGUIController when
tableChanged come in game.

Even if i catch the ArrayIndexOutOfBoundsException by the focusGained
Method inside the DataBaseGUIController i wrote some instructions when
the searchText object gain the focus we call the fundsachenTable and
getSelectionModel to clearSelection. all this help us to remove the
selection of the row