changes in the implementation of the methods for searching the database
LogicModelFundSachen: I have removed the checkDateFormat Method from
this class. Now this method do not exists anymore with this name. I have
declared one method with the name checkDateFormatToSearchInDataBase in
the LogicModel Super Class. So any LogicModel extended Class that will
search in DataBase by Date can access to this Method to check first the
Date Format before search in DataBase.

LogicModelFundSachen: now Override the Method searchResultsInDataBase
from Super Class. This searchResultsInDataBase Method is now declared in
the Super Class so any LogicModel extended class can access to this
Method to search in DataBase. And now by the conditional block switch
will be evaluate the return from the call to the new Method
getSearchSelected located in LogicModel Super Class.

LogicModel Super Class: I have declared all the variables to display
Error Message in Case the Date Format is not correct like Regex. Was
also declared the DateTimeFormatter, LocalDate and Date variables for
the Date we want to search in DataBase.

LogicModel: new Methods searchResultsInDataBase, getSearchSelected,
getToSearch, checkDateFormatToSearchInDataBase,
getDateToSearchInDataBase

DataBaseGUIController: by the keyPressed Method implements some new codes to evaluate if fundSachenTable is visible or Parking and depending of the results we call the searchResultsInDataBase Method.