New Method addNewParkingReservationToDataBase Parking Reservations.
LogicModelNewParking: new Method to prepared the Data we want to add to
the TAble Parking in Database. We create the ParkingReservation Object
and set all the values inside this method and call the existing
DAOParking object from this Method to add the new Reservation to the
Table Parking in the Database.

For adding the new reservation and with the help of the DAO Object we
call the method addNewParkingReservation.

additionally changes in this class i have removed unnecessary comments
and add some new instructions by the Method checkAnreiseDate and
checkAbreiseDate to set the Date value for the Arrival and Departure
Date values that will be saved in Database.