Removed the option to be able to delete entries from PARKING Table.
Customers who come to book a parking place directly do not have to be
deleted that entry from the database as well as customers who book
online with a third-party provider.

The parking ID for each entry is unique. It is generated by taking the
current year, counting how many entries we have in the table and adding
one. That is why we encounter the problem that if we delete any entry,
for the next entry there would be a duplicate parking ID.

We are not talking here about the unique ID of each entry within the
PARKING table in the database. We are talking only about the ID-Parking
column. this should be unique.

Changes Classes:

- DAOParking removed deleteParkingReservation Method.

- DaoParkingImpl removed the override deleteParkingReservation Method.

- DataBaseGUI removed the parkingTable.setComponentPopupMenu instruction

- DataBaseGUIController removed the instructions by actionPerformed
Method if databaseGUI.deleteItem (Instructions for the parkingTable so
it won't be displayed any more popupmenu in this JTable). It was also
removed by the Method popupMenuWillBecomeVisible all instructions that
has something to do with the parkingJTable.

- LogicModel removed all the instruction that has something to do with
deleting entries from parkingJTable.